Performed minor adjustments
